python/private_key/requirements.txt
py-clob-client-v2
python-dotenv
web3
requests
